+++ src/app/(main)/settings/hooks/useCategory.tsx
@@ -1,6 +1,7 @@
 import { Icon } from '@lobehub/ui';
 import { Tag } from 'antd';
-import { Bot, Brain, Cloudy, Info, Mic2, Settings2 } from 'lucide-react';
+import { Bot, Brain, Cloudy, Info, Mic2, Settings2, Sparkles } from 'lucide-react';
+import Link from 'next/link';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Flexbox } from 'react-layout-kit';
@@ -19,35 +20,71 @@ export const useCategory = () => {
         {
           icon: <Icon icon={Settings2} />,
           key: SettingsTabs.Common,
-          label: t('tab.common'),
+          label: (
+            <Link href={'/settings/common'} onClick={(e) => e.preventDefault()}>
+              {t('tab.common')}
+            </Link>
+          ),
+        },
+        {
+          icon: <Icon icon={Sparkles} />,
+          key: SettingsTabs.SystemAgent,
+          label: (
+            <Link href={'/settings/system-agent'} onClick={(e) => e.preventDefault()}>
+              {t('tab.system-agent')}
+            </Link>
+          ),
         },
         enableWebrtc && {
           icon: <Icon icon={Cloudy} />,
           key: SettingsTabs.Sync,
           label: (
-            <Flexbox align={'center'} gap={8} horizontal>
-              {t('tab.sync')}
-              <Tag bordered={false} color={'warning'}>
-                {t('tab.experiment')}
-              </Tag>
-            </Flexbox>
+            <Link href={'/settings/sync'} onClick={(e) => e.preventDefault()}>
+              <Flexbox align={'center'} gap={8} horizontal>
+                {t('tab.sync')}
+                <Tag bordered={false} color={'warning'}>
+                  {t('tab.experiment')}
+                </Tag>
+              </Flexbox>
+            </Link>
           ),
         },
         showLLM && {
           icon: <Icon icon={Brain} />,
           key: SettingsTabs.LLM,
-          label: t('tab.llm'),
+          label: (
+            <Link href={'/settings/llm'} onClick={(e) => e.preventDefault()}>
+              {t('tab.llm')}
+            </Link>
+          ),
         },
-        { icon: <Icon icon={Mic2} />, key: SettingsTabs.TTS, label: t('tab.tts') },
+
+        {
+          icon: <Icon icon={Mic2} />,
+          key: SettingsTabs.TTS,
+          label: (
+            <Link href={'/settings/tts'} onClick={(e) => e.preventDefault()}>
+              {t('tab.tts')}
+            </Link>
+          ),
+        },
         {
           icon: <Icon icon={Bot} />,
           key: SettingsTabs.Agent,
-          label: t('tab.agent'),
+          label: (
+            <Link href={'/settings/agent'} onClick={(e) => e.preventDefault()}>
+              {t('tab.agent')}
+            </Link>
+          ),
         },
         {
           icon: <Icon icon={Info} />,
           key: SettingsTabs.About,
-          label: t('tab.about'),
+          label: (
+            <Link href={'/settings/about'} onClick={(e) => e.preventDefault()}>
+              {t('tab.about')}
+            </Link>
+          ),
         },
       ].filter(Boolean) as MenuProps['items'],
     [t, enableWebrtc, showLLM],