+++ src/features/ModelSwitchPanel/index.tsx
@@ -1,6 +1,7 @@
 import { Icon } from '@lobehub/ui';
 import { Dropdown } from 'antd';
 import { createStyles } from 'antd-style';
+import type { ItemType } from 'antd/es/menu/interface';
 import isEqual from 'fast-deep-equal';
 import { LucideArrowRight } from 'lucide-react';
 import { useRouter } from 'next/navigation';
@@ -37,21 +38,24 @@ const useStyles = createStyles(({ css, prefixCls }) => ({
   `,
 }));
 
+const menuKey = (provider: string, model: string) => `${provider}-${model}`;
+
 const ModelSwitchPanel = memo<PropsWithChildren>(({ children }) => {
   const { t } = useTranslation('components');
   const { styles, theme } = useStyles();
-  const [model, updateAgentConfig] = useAgentStore((s) => [
+  const [model, provider, updateAgentConfig] = useAgentStore((s) => [
     agentSelectors.currentAgentModel(s),
+    agentSelectors.currentAgentModelProvider(s),
     s.updateAgentConfig,
   ]);
 
   const router = useRouter();
   const enabledList = useUserStore(modelProviderSelectors.modelProviderListForModelSelect, isEqual);
 
-  const items = useMemo(() => {
+  const items = useMemo<ItemType[]>(() => {
     const getModelItems = (provider: ModelProviderCard) => {
       const items = provider.chatModels.map((model) => ({
-        key: model.id,
+        key: menuKey(provider.id, model.id),
         label: <ModelItemRender {...model} />,
         onClick: () => {
           updateAgentConfig({ model: model.id, provider: provider.id });
@@ -90,7 +94,7 @@ const ModelSwitchPanel = memo<PropsWithChildren>(({ children }) => {
   return (
     <Dropdown
       menu={{
-        activeKey: model,
+        activeKey: menuKey(provider, model),
         className: styles.menu,
         items,
         style: {