+++ src/features/User/UserPanel/useNewVersion.tsx
@@ -1,4 +1,5 @@
 import { useGlobalStore } from '@/store/global';
+import { featureFlagsSelectors, useServerConfigStore } from '@/store/serverConfig';
 
 export const useNewVersion = () => {
   const [hasNewVersion, useCheckLatestVersion] = useGlobalStore((s) => [
@@ -6,7 +7,8 @@ export const useNewVersion = () => {
     s.useCheckLatestVersion,
   ]);
 
-  useCheckLatestVersion();
+  const { enableCheckUpdates } = useServerConfigStore(featureFlagsSelectors);
+  useCheckLatestVersion(enableCheckUpdates);
 
   return hasNewVersion;
 };